Shared Preferences (Bonus)...
Create a new project with two Edit Texts and two Buttons.
Name the Buttons 'sharing thru intent' and 'save to SHPFS'.

When the 'sharing thru intent' Button is clicked, get the data from the Edit Texts and display it in a Text View.
When the 'save to SHPFS' Button is clicked simply save the data to Shared Preferences.
In the next Activity, have a single Button that gets the data from Shared Preferences and displays it in a Text View when clicked.